killer done, need handle data-race

+++ philo/include/philosophers.h
@@ -6,7 +6,7 @@
 /*   By: Vitor <vsergio@student.42.rio>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/10 23:08:40 by Vitor             #+#    #+#             */
-/*   Updated: 2022/10/21 01:05:14 by Vitor            ###   ########.fr       */
+/*   Updated: 2022/10/21 13:51:43 by Vitor            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,16 +28,16 @@ typedef struct	s_data
 	long int			time_to_eat;
 	long int			*last_meal;
 	int					pos;
-}				t_data;
+}						t_data;
 
 typedef struct	s_philo
 {
 	pthread_t			philo_th;
 	t_data				data_control;
 	int					pos;
-}				t_philo;
+}						t_philo;
 
-long long int	get_current_time(t_data *data);
+long int	get_current_time(t_data *data);
 long int	ft_atoi(char *str);
 int			check_atoi_numbers(char *string);
 void		*lifetime(void *philo);
